PropertyConfigurerConfiguration factored out (shame I had to use inheritance instead of composition, but couldn't figure out how to)
