+++ mifosng-provider/src/test/java/org/mifosplatform/infrastructure/configuration/PropertyValueInjectTest.java
@@ -3,22 +3,23 @@ package org.mifosplatform.infrastructure.configuration;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mifosplatform.infrastructure.configuration.spring.PropertyConfigurerConfiguration;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.core.env.MutablePropertySources;
 import org.springframework.core.env.PropertySource;
 import org.springframework.mock.env.MockPropertySource;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import edu.emory.mathcs.backport.java.util.Collections;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration
 public class PropertyValueInjectTest {
 
     @Configuration
-    public static class PropertyValueInjectTestConfiguration {
+    public static class PropertyValueInjectTestConfiguration extends PropertyConfigurerConfiguration {
 
         @Bean
         ExampleTestConfigurableService exampleTestConfigurableService() {
@@ -32,22 +33,13 @@ public class PropertyValueInjectTest {
             return new ExampleTestDependendantService();
         }
 
-        @Bean
-        PropertySource<?> propertySource() {
+        @Override
+        @SuppressWarnings("unchecked")
+        protected Iterable<PropertySource<?>> getPropertySources() {
             MockPropertySource properties = new MockPropertySource();
             properties.setProperty("sample.string", "hello, world");
             properties.setProperty("sample.int", 123);
-            return properties;
-        }
-
-        @Bean
-        PropertySourcesPlaceholderConfigurer propertyConfigurer() {
-            // http://blog.jamesdbloom.com/UsingPropertySourceAndEnvironment.html
-            PropertySourcesPlaceholderConfigurer propertyConfigurer = new PropertySourcesPlaceholderConfigurer();
-            MutablePropertySources propertySources = new MutablePropertySources();
-            propertySources.addFirst(propertySource());
-            propertyConfigurer.setPropertySources(propertySources);
-            return propertyConfigurer;
+            return Collections.singleton(properties);
         }
     }
